+++ warp/examples/fem/example_burgers.py
@@ -18,7 +18,6 @@
 import warp as wp
 import warp.examples.fem.utils as fem_example_utils
 import warp.fem as fem
-import warp.sparse as sp
 
 
 @fem.integrand
@@ -141,19 +140,19 @@ class Example:
         matrix_inertia = fem.integrate(
             vel_mass_form, fields={"u": trial, "v": self._test}, output_dtype=wp.float32, nodal=True
         )
-        self._inv_mass_matrix = sp.bsr_copy(matrix_inertia)
+        self._inv_mass_matrix = wp.sparse.bsr_copy(matrix_inertia)
         fem_example_utils.invert_diagonal_bsr_matrix(self._inv_mass_matrix)
 
         # Initial condition
         self.velocity_field = vector_space.make_field()
         fem.interpolate(initial_condition, dest=self.velocity_field)
 
-        # Velocity nor field -- for visualization purposes
+        # Velocity norm field -- for visualization purposes
         self.velocity_norm_field = scalar_space.make_field()
         fem.interpolate(velocity_norm, dest=self.velocity_norm_field, fields={"u": self.velocity_field})
 
         self.renderer = fem_example_utils.Plot()
-        self.renderer.add_surface("u_norm", self.velocity_norm_field)
+        self.renderer.add_field("u_norm", self.velocity_norm_field)
 
     def _velocity_delta(self, trial_velocity):
         # Integration on sides
@@ -178,7 +177,7 @@ class Example:
                 alpha=1.0,
                 beta=1.0,
             )
-        return sp.bsr_mv(self._inv_mass_matrix, rhs)
+        return self._inv_mass_matrix @ rhs
 
     def step(self):
         self.current_frame += 1
@@ -215,7 +214,7 @@ class Example:
 
     def render(self):
         self.renderer.begin_frame(time=self.current_frame * self.sim_dt)
-        self.renderer.add_surface("u_norm", self.velocity_norm_field)
+        self.renderer.add_field("u_norm", self.velocity_norm_field)
         self.renderer.end_frame()